working #time 1h #comment reorganizing controllers and routes to separate out auth
