+++ server/routes/api-routes/index.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const boxRoutes = require('./box-routes');
-const authRoutes = require('./auth-routes');
+const authRoutes = require('../auth-routes');
 
 router.use('/box', boxRoutes);
 router.use('/', authRoutes);
