+++ server/routes/auth-routes.js
@@ -1,12 +1,11 @@
 
 const router = require('express').Router();
-const { signup, login, callback, logout } = require("../../controllers/user-controller");
+const { signup, login, logout } = require("../controllers/auth-controller");
 
 require("dotenv").config();
 
 router.route('/signup').get(signup);
 router.route('/login').get(login);
-router.route('/callback').get(callback);
 router.route('/logout').get(logout);
 
 module.exports = router;
