+++ server/routes/index.js
@@ -1,8 +1,10 @@
 const router = require('express').Router();
-
+const authRoutes = require('./auth-routes');
 const apiRoutes = require('./api-routes');
 
-router.use('/', apiRoutes);
+// router.use('/', viewRoutes);
+router.use('/auth', authRoutes);
+router.use('/api', apiRoutes);
 
 router.use((req, res) => {
     res.send(req.oidc.isAuthenticated() ? 'Logged In' : 'Logged Out');
